Added common footer to every route
