+++ simplq/src/components/Layout.jsx
@@ -1,10 +1,12 @@
 import React from 'react';
 import Routes from './Routes';
+import { Footer } from './common/Footer'
 
 function Layout() {
   return (
     <>
       <Routes />
+      <Footer />
     </>
   );
 }
