+++ simplq/src/components/pages/Home/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { BenefitsInfo, HowToCreate, HowToJoin, ExtraInfo, Footer } from './StaticInfos';
+import { BenefitsInfo, HowToCreate, HowToJoin, ExtraInfo} from './StaticInfos';
 import LandingPage from './LandingPage';
 import Navbar from './Nav/Navbar';
 
@@ -12,7 +12,6 @@ const Home = (props) => {
       <HowToCreate />
       <HowToJoin />
       <ExtraInfo />
-      <Footer />
     </>
   );
 };
